requirements.txt
# Core dependencies
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Environment
python-dotenv>=1.0.0

# Financial data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0

# Web search
duckduckgo-search>=6.0.0

# News/RSS
feedparser>=6.0.0

# Testing
pytest>=8.0.0
